32x: use byte[] for cache data, update lib

+++ src/main/java/sh2/Sh2Memory.java
@@ -144,7 +144,7 @@ public final class Sh2Memory implements IMemory {
 				if (true) throw new RuntimeException();
 				break;
 		}
-		return (int) (res & size.getMask());
+		return res & size.getMask();
 	}
 
 	@Override
